+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.10.2.1-beta2 of SecurityCraft--------------------------
 
 - Fix: The button to confirm Secure Trading Station transactions is not interactable for other players
+- Fix: Passcode-protected blocks can be set up with an empty passcode
 
 --------------------------Changelog for v1.10.2.1-beta1 of SecurityCraft--------------------------